pyproject.toml
[tool.poetry]
name = "agent-cooper"
version = "0.1.0"
description = "Send out notifications automatically with AWS free services"
authors = ["404pilot <404pilot@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "^2021.1"
requests = "^2.26.0"
google-api-python-client = "^2.14.1"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
html2text = "^2020.1.16"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]
taskipy = "^1.1.2"
boto3 = "^1.18.6"
PyYAML = "^5.4.1"

[tool.poetry.scripts]

[tool.taskipy.tasks]
generate_lambda_requirement = """mkdir -p ./.build && poetry export --with-credentials --without-hashes -f requirements.txt > ./.build/lambda-layer-requirements.txt"""

generate_lambda_deps = """pip install --disable-pip-version-check -r ./.build/lambda-layer-requirements.txt -t ./.build/dependencies/python/lib/python3.8/site-packages"""

sync_secrets_properties = """
    AWS_PROFILE=${AWS_PROFILE} sops -d ./properties/conf.yaml \
        | python ./properties/sync_properties.py - \
            -r "${AWS_DEFAULT_REGION}" \
            -s "${STAGE}" \
            -p "${AWS_PROFILE}" \
            --project-name ${PROJECT_NAME} \
            --properties-type SecureString
"""

sam_clean = "rm -rf .aws-sam"

s3_prep = """
    aws s3api head-bucket \
        --profile "${AWS_PROFILE}" \
        --bucket "${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        --no-cli-pager \
        || ( aws s3 mb \
        --profile "${AWS_PROFILE}" \
        --region "${AWS_DEFAULT_REGION}" \
        --no-cli-pager \
        "s3://${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        && aws s3api put-bucket-tagging \
        --profile "${AWS_PROFILE}" \
        --region "${AWS_DEFAULT_REGION}" \
        --bucket "${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        --tagging "TagSet=[{Key=Project,Value=${PROJECT_NAME}}]" \
        --no-cli-pager )
"""

sam_build = """
    sam build \
        --debug \
        --template-file ./infra/template.yaml \
        --profile "${AWS_PROFILE}" \
        --region "${AWS_DEFAULT_REGION}" \
        --parameter-overrides "\
            ProjectName=${PROJECT_NAME} \
            Stage=${STAGE} \
        "
"""

sam_deploy = """
    sam deploy \
        --debug \
        --stack-name "${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        --s3-bucket "${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        --s3-prefix "infra/project" \
        --capabilities CAPABILITY_AUTO_EXPAND CAPABILITY_NAMED_IAM CAPABILITY_IAM \
        --region "${AWS_DEFAULT_REGION}" \
        --profile "${AWS_PROFILE}" \
        --parameter-overrides "\
            ProjectName=${PROJECT_NAME} \
            Stage=${STAGE} \
        " \
        --tags Project=${PROJECT_NAME} \
        --no-fail-on-empty-changeset
"""

aws_delete_stack = """
    aws cloudformation delete-stack \
        --debug \
        --stack-name "${PROJECT_NAME}-${STAGE}-${AWS_DEFAULT_REGION}" \
        --region "${AWS_DEFAULT_REGION}" \
        --profile "${AWS_PROFILE}" \
        --no-cli-pager
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
